Small clean up on chess board, add min max ai, better perft testing

+++ source/tools/testing.h
@@ -7,6 +7,6 @@
 const std::string DEFAULT_POSITION = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1";
 
 namespace testing {
-	unsigned long long perft(const std::string& fen, uint64_t startingDepth, bool verbose);
-	unsigned long long perft(uint64_t startingDepth);
+	uint64_t perft(const std::string& fen, uint8_t startingDepth, bool verbose);
+	uint64_t perft(uint8_t startingDepth);
 }
